+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Syscall stub function (credit @b33f)
 - Added SharpSploit.Execution.Injection namespace (credit @TheWover)
 - Added SharpSploit.Pivoting namespace, reverse port forwarding (credit @rasta-mouse)
+- Added error/verbose output to PowerShellExecute function
 
 ## [v1.5] - 2019-12-27
 ### Added